Added beta implementation of filling fields in client_edit html page
