--- invoicer_project/clients/templates/menu_wrapper.html
@@ -1,34 +1,0 @@
-{% load static %}
-<div id="sidebar-wrapper">
-   <ul class="sidebar-nav nav-pills nav-stacked" id="menu">
-      <li>
-         <a href="#"><span class="fa-stack fa-lg pull-left invoices_icon"><img class="icon_menu"
-            src="{% static 'img/invoice.png' %}"/></span>Invoices</a>
-      </li>
-      <li class="active">
-         <a href="{% url 'home' %}"
-            ><span class="fa-stack fa-lg pull-left clients_icon"
-            ><img class="icon_menu" src="{% static 'img/clients.png' %}"/></span
-            >Clients</a
-            >
-      </li>
-      <li>
-         <a href="#"
-            ><span class="fa-stack fa-lg pull-left"
-            ><img class="icon_menu" src="{% static 'img/list-items.png' %}"/></span
-            >Items</a
-            >
-      </li>
-      <li>
-         <a href="#">
-         <span class="fa-stack fa-lg pull-left"
-            ><img class="icon_menu" src="{% static 'img/statistics.png' %}"/></span
-            >Statistics</a
-            >
-      </li>
-      <li>
-         <a href="#"><span class="fa-stack fa-lg pull-left"><img class="icon_menu"
-            src="{% static 'img/settings.png' %}"/></span>Settings</a>
-      </li>
-   </ul>
-</div>
