+++ invoicer_project/static/js/clients_list_request.js
@@ -89,34 +89,17 @@ async function addElementsDynamically() {
     }
 }
 
-function getClientById(clientId) {
-    return fetch(host + '/clients/client_by_id/' + clientId, {
-        method: "GET",
-        headers: {
-            'Authorization': `Bearer ${window.localStorage.getItem('accessToken')}`
-        },
-    })
-        .then(response => response.json())
-        .catch(error => console.error(error));
-}
+
 
 function addEditButtonListeners() {
     const editButtons = document.querySelectorAll('.edit-client');
-    let responseFromServer, clientId;
+    let clientId;
     editButtons.forEach(span => {
         span.addEventListener('click', async () => {
             clientId = span.dataset.clientId;
-            responseFromServer = await getClientById(clientId);
-            console.log(responseFromServer);
-            // window.location.replace(host + '/clients/edit/' + clientId);
-            // <!-- ERROR set data to fields won't be executed -->
-            setDataToFields(responseFromServer)
+            window.location.href = host + "/clients/edit/" + clientId;
         });
     });
-}
-
-function setDataToFields(responseFromServer) {
-    document.getElementById("name_input_client_edit_page").value = responseFromServer['first_name'];
 }
 
 function addDeleteButtonListeners() {
